move `downloadWorkerConfig`, `mapBindings` and `assertNever` to utils
