+++ packages/wrangler/src/utils/assert-never.ts
@@ -1,0 +1,1 @@
+export function assertNever(_value: never) {}
